Lock down test gems' Gemfile.lock
